+++ ddth-akka-core/src/main/java/com/github/ddth/akka/scheduling/TickFanOutActor.java
@@ -1,32 +1,31 @@
 package com.github.ddth.akka.scheduling;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.concurrent.TimeUnit;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import com.github.ddth.commons.utils.MapUtils;
-
-import akka.ConfigurationException;
 import akka.actor.AbstractActor;
 import akka.actor.ActorSystem;
 import akka.actor.Cancellable;
-import scala.concurrent.ExecutionContextExecutor;
+import com.github.ddth.commons.utils.MapUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import scala.concurrent.duration.Duration;
+
+import java.util.concurrent.TimeUnit;
 
 /**
  * Actor that sends "tick" messages to all subscribed workers every "tick".
- * 
+ *
  * <p>
  * After {@link ActorSystem} is built, create one instance of {@link TickFanOutActor} to broadcast "tick" messages.
  * </p>
- * 
+ *
  * @author Thanh Nguyen <btnguyen2k@gmail.com>
  * @since 0.1.0
  */
 public abstract class TickFanOutActor extends AbstractActor {
+    /**
+     * This tag is attached to the tick-message, containing the path-string of send (i.e. the tick-fanout actor).
+     */
+    public final static String TAG_SENDDER_ADDR = "sender_addr";
+
     /**
      * Marker message to signal {@link TickFanOutActor} that a "tick" has come.
      */
@@ -43,7 +42,7 @@ public abstract class TickFanOutActor extends AbstractActor {
      * (https://doc.akka.io/docs/akka/2.5/scheduler.html), Akka scheduler is not
      * designed for long-term scheduling. Hence, we renew the clock every 24g,
      * for now.
-     * 
+     *
      * @since 0.1.3
      */
     protected void renewClock() {
@@ -56,23 +55,19 @@ public abstract class TickFanOutActor extends AbstractActor {
 
     /**
      * Start the "clock" to send "tick" every second.
-     * 
+     *
      * @since 0.1.3
      */
     protected void startClock() {
-        clock = getContext().system().scheduler().schedule(Duration.create(0, TimeUnit.SECONDS),
-                Duration.create(1, TimeUnit.SECONDS), () -> {
-                    self().tell(new OnTick(), self());
-                }, getContext().dispatcher());
+        clock = getContext().system().scheduler()
+                .schedule(Duration.create(0, TimeUnit.SECONDS), Duration.create(1, TimeUnit.SECONDS),
+                        () -> self().tell(new OnTick(), self()), getContext().dispatcher());
         timestampClockStarted = System.currentTimeMillis();
-        // getTimers().startPeriodicTimer(this.getClass().getSimpleName(), new
-        // OnTick(),
-        // Duration.create(1, TimeUnit.SECONDS));
     }
 
     /**
      * Stop the "clock" that sends "tick" every second.
-     * 
+     *
      * @since 0.1.3
      */
     protected void stopClock() {
@@ -80,7 +75,6 @@ public abstract class TickFanOutActor extends AbstractActor {
             if (clock != null) {
                 clock.cancel();
             }
-            // getTimers().cancel(this.getClass().getSimpleName());
         } catch (Exception e) {
             LOGGER.warn(e.getMessage(), e);
         } finally {
@@ -112,13 +106,14 @@ public abstract class TickFanOutActor extends AbstractActor {
     @Override
     public Receive createReceive() {
         return receiveBuilder().match(OnTick.class, tick -> {
-            fanOut(new TickMessage(MapUtils.createMap("sender", self().path().toString())));
+            fanOut(new TickMessage(MapUtils.createMap(TAG_SENDDER_ADDR, self().path().toString())));
+            renewClock();
         }).matchAny(this::onReceive).build();
     }
 
     /**
      * This method is called to handle incoming message.
-     * 
+     *
      * @param message
      */
     protected void onReceive(Object message) {
@@ -135,30 +130,10 @@ public abstract class TickFanOutActor extends AbstractActor {
 
     /**
      * Convenient method to get the associated actor system.
-     * 
+     *
      * @return
      */
     protected ActorSystem getActorSystem() {
         return context().system();
-    }
-
-    protected static Map<String, Boolean> exceptionLoggedGetECE = new HashMap<>();
-
-    /**
-     * Get the {@link ExecutionContextExecutor} instance to do async work.
-     *
-     * @param name
-     * @return
-     */
-    protected ExecutionContextExecutor getExecutionContextExecutor(String name) {
-        try {
-            return getActorSystem().dispatchers().lookup(name);
-        } catch (ConfigurationException e) {
-            if (exceptionLoggedGetECE.get(name) == null) {
-                LOGGER.warn(e.getMessage());
-                exceptionLoggedGetECE.put(name, Boolean.TRUE);
-            }
-            return getActorSystem().dispatcher();
-        }
     }
 }