+++ ddth-akka-core/src/test/java/com/github/ddth/akka/qnd/QndMultiNodeTickFanOutPubSub.java
@@ -5,6 +5,7 @@ import akka.actor.ActorSystem;
 import akka.actor.Props;
 import com.github.ddth.akka.AkkaUtils;
 import com.github.ddth.akka.scheduling.BaseWorker;
+import com.github.ddth.akka.scheduling.TickFanOutActor;
 import com.github.ddth.akka.scheduling.TickMessage;
 import com.github.ddth.akka.scheduling.annotation.Scheduling;
 import com.github.ddth.akka.scheduling.tickfanout.MultiNodePubSubBasedTickFanOutActor;
@@ -24,27 +25,39 @@ public class QndMultiNodeTickFanOutPubSub {
         System.setProperty("org.slf4j.simpleLogger.showShortLogName", "false");
     }
 
+    private final static String DF = "HH:mm:ss.SSS";
+
     @Scheduling("*/3 * *")
     private static class MyWorker1 extends BaseWorker {
+        public MyWorker1() {
+            setHandleMessageAsync(true);
+        }
+
         @Override
         protected void doJob(String lockId, TickMessage tick) {
             Date now = new Date();
-            System.out.println("{" + self().path() + "}: Tick " + tick.getId() + " / Now " + DateFormatUtils
-                    .toString(now, DateFormatUtils.DF_ISO8601) + " / TickTime " + DateFormatUtils
-                    .toString(tick.getTimestamp(), DateFormatUtils.DF_ISO8601) + " / Lag " + (now.getTime() - tick
-                    .getTimestamp().getTime()));
+            System.out.println(
+                    "{" + self().path().name() + "}: Tick {" + tick.getId() + "} from {" + sender().path().name()
+                            + " : " + tick.getTag(TickFanOutActor.TAG_SENDDER_ADDR) + "} / Now " + DateFormatUtils
+                            .toString(now, DF) + " / TickTime " + DateFormatUtils.toString(tick.getTimestamp(), DF)
+                            + " / Lag " + (now.getTime() - tick.getTimestamp().getTime()));
         }
     }
 
     @Scheduling("*/5 * *")
     private static class MyWorker2 extends BaseWorker {
+        public MyWorker2() {
+            setHandleMessageAsync(false);
+        }
+
         @Override
         protected void doJob(String lockId, TickMessage tick) {
             Date now = new Date();
-            System.out.println("{" + self().path() + "}: Tick " + tick.getId() + " / Now " + DateFormatUtils
-                    .toString(now, DateFormatUtils.DF_ISO8601) + " / TickTime " + DateFormatUtils
-                    .toString(tick.getTimestamp(), DateFormatUtils.DF_ISO8601) + " / Lag " + (now.getTime() - tick
-                    .getTimestamp().getTime()));
+            System.out.println(
+                    "{" + self().path().name() + "}: Tick {" + tick.getId() + "} from {" + sender().path().name()
+                            + " : " + tick.getTag(TickFanOutActor.TAG_SENDDER_ADDR) + "} / Now " + DateFormatUtils
+                            .toString(now, DF) + " / TickTime " + DateFormatUtils.toString(tick.getTimestamp(), DF)
+                            + " / Lag " + (now.getTime() - tick.getTimestamp().getTime()));
         }
     }